requirements.txt
streamlit
pymongo
python-dotenv
fastapi
uvicorn
SpeechRecognition
pyaudio
pydub
python-multipart
requests
numpy
pocketsphinx 